remove flash_ helper
* this was confusing for implementations because it wasn't transparent
what the underlying key was. Especially in the case of using the
whitelist feature for Delete sessions, because it wasn't obvious you'd
have to pass in something like flash_flash_success instead of just
flash_success. There's a risk of naming collisions now, implementations
will just have to be careful.